Don't require lambda.json
Remove the lambda.json requirement. It breaks too many things on different platforms, and we don't really need it. Just specify the ignored files directly.

Aside -- we also added a conditional to avoid fetching AWS account IDs if we don't need them. This made it easier to test the above feature, and saved some steps in the algorithm as well.

Fixes #65.