Introduce timing and iterative snapshot cleanup
- Add the ability to clean up snapshots iteratively:

We now pull a few tag values at a time for the 'DeleteOn' tag, process those deletes, and then retry. Coupled with the item below, this means we're able to do more work, without erroring as often. This means we'll now be able to cleanup snapshots of any "DeleteOn" value, instead of only values in the last 10 days.

- Add timing measurements on cleanup:

We now regularly check if we've exceeded 4 minutes, in between each snapshot date cleanup, and if we do, we bail out and return success. This means we will iteratively cleanup snapshots, but won't blow up if it takes a while.

Fixes #66.