Be sure SNS methods take region arg
Be sure all SNS methods are passed a region argument. We do this to avoid getting the default from the environment, which may not be `us-east-1`. We require the tool be installed there, but this is tricky since the non-fanout jobs don't get called by the CLI, so they can never override the installed region. Maybe we should ditch the idea that it can be installed in any region except us-east-1?

Fixes #83.