Cut away the 'every DeleteOn tag' strategy
A while back, we moved to a strategy where we looked up all snapshots with a DeleteOn tag, and cleaned them if needed. There's just _so many_ snapshots that we processed this way, even ones more than a year old, that we could never clean up; we looked at them too.

This reverts that a bit, with a more reasonable '7 days back' strategy. We still delete anything that should have expired in the last 7 days, regardless of when it was created; as long as this runs every 7 days, we don't need to worry about missing a snapshot that should have been deleted.