Add option to ignore retention settings
Normally, when cleaning up snapshots, this tool cares about how many snapshots of a volume remain, and as well as other things like being able to map a snapshot back to an existing instance. This adds a top-level flag `ignore_retention` that simply causes the tool to delete all snapshots with a `DeleteOn` tag value that's older than today (8 days back currently is the max).

This change, when the flag is set, effectively cleans up snapshots from autoscale groups. Note that the flag affects behavior globally -- when we stop considering where a snapshot came from, cleanup becomes an account-wide process.

Fixes #14.