Only use ignore_retention if necessary
Make a slight behavior adjustment to the ignore_retention flag. Only ignore retention, when cleaning up snapshots, if we can't find an instance and volume to map a snapshot back to. We would normally use the mapping to count how many other snapshots exist. If we simply fail to do that, don't assume the count is zero either -- use the new flag.

Fixes #14.